+++ foundation/objectmap.h
@@ -16,18 +16,18 @@
 #include <foundation/types.h>
 
 
-FOUNDATION_API objectmap_t*     objectmap_allocate( int size );
+FOUNDATION_API objectmap_t*     objectmap_allocate( size_t size );
 FOUNDATION_API void             objectmap_deallocate( objectmap_t* map );
 
-FOUNDATION_API void             objectmap_initialize( objectmap_t* map, int size );
+FOUNDATION_API void             objectmap_initialize( objectmap_t* map, size_t size );
 FOUNDATION_API void             objectmap_finalize( objectmap_t* map );
 
-FOUNDATION_API int              objectmap_size( const objectmap_t* map );
+FOUNDATION_API size_t           objectmap_size( const objectmap_t* map );
 FOUNDATION_API object_t         objectmap_reserve( objectmap_t* map );
 FOUNDATION_API void             objectmap_free( objectmap_t* map, object_t id );
 FOUNDATION_API void             objectmap_set( objectmap_t* map, object_t id, void* object );
 
-FOUNDATION_API void*            objectmap_raw_lookup( const objectmap_t* map, int index );
+FOUNDATION_API void*            objectmap_raw_lookup( const objectmap_t* map, size_t index );
 FOUNDATION_API void*            objectmap_lookup_ref( const objectmap_t* map, object_t id );
 FOUNDATION_API bool             objectmap_lookup_unref( const objectmap_t* map, object_t id, object_deallocate_fn deallocate );
