+++ foundation/process.h
@@ -24,7 +24,7 @@ FOUNDATION_API void        process_finalize( process_t* proc );
 
 FOUNDATION_API void        process_set_working_directory( process_t* proc, const char* path );
 FOUNDATION_API void        process_set_executable_path( process_t* proc, const char* path );
-FOUNDATION_API void        process_set_arguments( process_t* proc, const char** args, unsigned int num );
+FOUNDATION_API void        process_set_arguments( process_t* proc, const char** args, size_t num );
 FOUNDATION_API void        process_set_flags( process_t* proc, unsigned int flags );
 FOUNDATION_API void        process_set_verb( process_t* proc, const char* verb );
 FOUNDATION_API void        process_set_exit_code( int code );
