+++ foundation/regex.h
@@ -16,6 +16,6 @@
 
 
 FOUNDATION_API regex_t*  regex_compile( const char* pattern );
-FOUNDATION_API bool      regex_match( regex_t* regex, const char* input, int inlength, regex_capture_t* captures, int maxcaptures );
+FOUNDATION_API bool      regex_match( regex_t* regex, const char* input, size_t inlength, regex_capture_t* captures, size_t maxcaptures );
 FOUNDATION_API void      regex_deallocate( regex_t* regex );
